+++ examples/emi-entity/sdk/Entity1Entity.go
@@ -4,6 +4,8 @@ import (
 	"encoding"
 	"encoding/json"
 	"github.com/torabian/emi/emigo"
+	"github.com/torabian/emi/emigorm"
+	"gorm.io/gorm"
 )
 
 // The base class definition for entity1Entity
@@ -743,4 +745,825 @@ func CastEntity1EntityNestedContainerOptNestedInnerNestedItemsOptFromCli(c emigo
 // TableName overrides the default gorm table name for Entity1Entity.
 func (x *Entity1Entity) TableName() string {
 	return "entity1_table"
+}
+
+// The base class definition for entity1EntityUpdateInput
+type Entity1EntityUpdateInput struct {
+	Title              emigo.Nullable[string]                                     `json:"title" yaml:"title"`
+	Items              emigo.ArrayNullable[Entity1EntityItems]                    `json:"items" yaml:"items"`
+	Items2             emigo.ArrayNullable[Entity1EntityItems2]                   `json:"items2" yaml:"items2"`
+	Items3             emigo.CollectionNullable[Entity2Entity]                    `json:"items3" yaml:"items3"`
+	Items4             emigo.CollectionNullable[Entity2Entity]                    `json:"items4" yaml:"items4"`
+	Owner              emigo.OneNullable[Entity2Entity]                           `json:"owner" yaml:"owner"`
+	Manager            emigo.OneNullable[Entity2Entity]                           `json:"manager" yaml:"manager"`
+	Content1           emigo.Nullable[Entity1EntityUpdateInputContent1]           `json:"content1" yaml:"content1"`
+	Content2           emigo.Nullable[Entity1EntityUpdateInputContent2]           `json:"content2" yaml:"content2"`
+	Complex1           *Money                                                     `json:"complex1" yaml:"complex1"`
+	Subtitle           emigo.Nullable[string]                                     `json:"subtitle" yaml:"subtitle"`
+	IsActive           emigo.Nullable[bool]                                       `json:"isActive" yaml:"isActive"`
+	IsFeatured         emigo.Nullable[bool]                                       `json:"isFeatured" yaml:"isFeatured"`
+	ViewCount          emigo.Nullable[int]                                        `json:"viewCount" yaml:"viewCount"`
+	ViewCountOpt       emigo.Nullable[int]                                        `json:"viewCountOpt" yaml:"viewCountOpt"`
+	SmallCount         emigo.Nullable[int32]                                      `json:"smallCount" yaml:"smallCount"`
+	SmallCountOpt      emigo.Nullable[int32]                                      `json:"smallCountOpt" yaml:"smallCountOpt"`
+	BigCount           emigo.Nullable[int64]                                      `json:"bigCount" yaml:"bigCount"`
+	BigCountOpt        emigo.Nullable[int64]                                      `json:"bigCountOpt" yaml:"bigCountOpt"`
+	Ratio32            emigo.Nullable[float32]                                    `json:"ratio32" yaml:"ratio32"`
+	Ratio32Opt         emigo.Nullable[float32]                                    `json:"ratio32Opt" yaml:"ratio32Opt"`
+	Ratio64            emigo.Nullable[float64]                                    `json:"ratio64" yaml:"ratio64"`
+	Ratio64Opt         emigo.Nullable[float64]                                    `json:"ratio64Opt" yaml:"ratio64Opt"`
+	Status             emigo.Nullable[string]                                     `json:"status" yaml:"status"`
+	StatusOpt          emigo.Nullable[string]                                     `json:"statusOpt" yaml:"statusOpt"`
+	Metadata           emigo.Nullable[map[string]string]                          `json:"metadata" yaml:"metadata"`
+	MetadataOpt        emigo.Nullable[map[string]string]                          `json:"metadataOpt" yaml:"metadataOpt"`
+	RawSettings        emigo.Nullable[map[string]string]                          `json:"rawSettings" yaml:"rawSettings"`
+	Labels             emigo.Nullable[[]string]                                   `json:"labels" yaml:"labels"`
+	LabelsOpt          emigo.Nullable[[]string]                                   `json:"labelsOpt" yaml:"labelsOpt"`
+	Misc               interface{}                                                `json:"misc" yaml:"misc"`
+	NestedContainer    emigo.Nullable[Entity1EntityUpdateInputNestedContainer]    `json:"nestedContainer" yaml:"nestedContainer"`
+	NestedContainerOpt emigo.Nullable[Entity1EntityUpdateInputNestedContainerOpt] `json:"nestedContainerOpt" yaml:"nestedContainerOpt"`
+}
+
+// The base class definition for content1
+type Entity1EntityUpdateInputContent1 struct {
+	Item1 emigo.Nullable[int64] `json:"item1" yaml:"item1"`
+}
+
+// The base class definition for content2
+type Entity1EntityUpdateInputContent2 struct {
+	Item2 emigo.Nullable[int64] `json:"item2" yaml:"item2"`
+}
+
+// The base class definition for nestedContainer
+type Entity1EntityUpdateInputNestedContainer struct {
+	NestedInner emigo.Nullable[Entity1EntityUpdateInputNestedContainerNestedInner] `json:"nestedInner" yaml:"nestedInner"`
+}
+
+// The base class definition for nestedInner
+type Entity1EntityUpdateInputNestedContainerNestedInner struct {
+	NestedItems emigo.ArrayNullable[Entity1EntityNestedContainerNestedInnerNestedItems] `json:"nestedItems" yaml:"nestedItems"`
+	NestedOwner emigo.OneNullable[Entity2Entity]                                        `json:"nestedOwner" yaml:"nestedOwner"`
+}
+
+// The base class definition for nestedContainerOpt
+type Entity1EntityUpdateInputNestedContainerOpt struct {
+	NestedInner emigo.Nullable[Entity1EntityUpdateInputNestedContainerOptNestedInner] `json:"nestedInner" yaml:"nestedInner"`
+}
+
+// The base class definition for nestedInner
+type Entity1EntityUpdateInputNestedContainerOptNestedInner struct {
+	NestedItemsOpt emigo.ArrayNullable[Entity1EntityNestedContainerOptNestedInnerNestedItemsOpt] `json:"nestedItemsOpt" yaml:"nestedItemsOpt"`
+}
+
+func (x *Entity1EntityUpdateInput) Json() string {
+	if x != nil {
+		str, _ := json.MarshalIndent(x, "", "  ")
+		return string(str)
+	}
+	return ""
+}
+func GetEntity1EntityUpdateInputCliFlags(prefix string) []emigo.CliFlag {
+	return []emigo.CliFlag{
+		{
+			Name: prefix + "title",
+			Type: "string?",
+		},
+		{
+			Name: prefix + "items",
+			Type: "complex",
+		},
+		{
+			Name: prefix + "items2",
+			Type: "complex",
+		},
+		{
+			Name: prefix + "items3",
+			Type: "complex",
+		},
+		{
+			Name: prefix + "items4",
+			Type: "complex",
+		},
+		{
+			Name: prefix + "owner",
+			Type: "complex",
+		},
+		{
+			Name: prefix + "manager",
+			Type: "complex",
+		},
+		{
+			Name: prefix + "content1",
+			Type: "object?",
+		},
+		{
+			Name: prefix + "content2",
+			Type: "object?",
+		},
+		{
+			Name: prefix + "complex1",
+			Type: "complex",
+		},
+		{
+			Name: prefix + "subtitle",
+			Type: "string?",
+		},
+		{
+			Name: prefix + "is-active",
+			Type: "bool?",
+		},
+		{
+			Name: prefix + "is-featured",
+			Type: "bool?",
+		},
+		{
+			Name: prefix + "view-count",
+			Type: "int?",
+		},
+		{
+			Name: prefix + "view-count-opt",
+			Type: "int?",
+		},
+		{
+			Name: prefix + "small-count",
+			Type: "int32?",
+		},
+		{
+			Name: prefix + "small-count-opt",
+			Type: "int32?",
+		},
+		{
+			Name: prefix + "big-count",
+			Type: "int64?",
+		},
+		{
+			Name: prefix + "big-count-opt",
+			Type: "int64?",
+		},
+		{
+			Name: prefix + "ratio32",
+			Type: "float32?",
+		},
+		{
+			Name: prefix + "ratio32-opt",
+			Type: "float32?",
+		},
+		{
+			Name: prefix + "ratio64",
+			Type: "float64?",
+		},
+		{
+			Name: prefix + "ratio64-opt",
+			Type: "float64?",
+		},
+		{
+			Name: prefix + "status",
+			Type: "enum?",
+		},
+		{
+			Name: prefix + "status-opt",
+			Type: "enum?",
+		},
+		{
+			Name: prefix + "metadata",
+			Type: "map?",
+		},
+		{
+			Name: prefix + "metadata-opt",
+			Type: "map?",
+		},
+		{
+			Name: prefix + "raw-settings",
+			Type: "map?",
+		},
+		{
+			Name: prefix + "labels",
+			Type: "slice?",
+		},
+		{
+			Name: prefix + "labels-opt",
+			Type: "slice?",
+		},
+		{
+			Name: prefix + "misc",
+			Type: "any",
+		},
+		{
+			Name: prefix + "nested-container",
+			Type: "object?",
+		},
+		{
+			Name: prefix + "nested-container-opt",
+			Type: "object?",
+		},
+	}
+}
+func CastEntity1EntityUpdateInputFromCli(c emigo.CliCastable) Entity1EntityUpdateInput {
+	data := Entity1EntityUpdateInput{}
+	if c.IsSet("title") {
+		emigo.ParseNullable(c.String("title"), &data.Title)
+	}
+	if c.IsSet("items") {
+		if u, ok := any(&data.Items).(encoding.TextUnmarshaler); ok {
+			u.UnmarshalText([]byte(c.String("items")))
+		}
+	}
+	if c.IsSet("items2") {
+		if u, ok := any(&data.Items2).(encoding.TextUnmarshaler); ok {
+			u.UnmarshalText([]byte(c.String("items2")))
+		}
+	}
+	if c.IsSet("items3") {
+		if u, ok := any(&data.Items3).(encoding.TextUnmarshaler); ok {
+			u.UnmarshalText([]byte(c.String("items3")))
+		}
+	}
+	if c.IsSet("items4") {
+		if u, ok := any(&data.Items4).(encoding.TextUnmarshaler); ok {
+			u.UnmarshalText([]byte(c.String("items4")))
+		}
+	}
+	if c.IsSet("owner") {
+		if u, ok := any(&data.Owner).(encoding.TextUnmarshaler); ok {
+			u.UnmarshalText([]byte(c.String("owner")))
+		}
+	}
+	if c.IsSet("manager") {
+		if u, ok := any(&data.Manager).(encoding.TextUnmarshaler); ok {
+			u.UnmarshalText([]byte(c.String("manager")))
+		}
+	}
+	if c.IsSet("content1") {
+		emigo.ParseNullable(c.String("content1"), &data.Content1)
+	}
+	if c.IsSet("content2") {
+		emigo.ParseNullable(c.String("content2"), &data.Content2)
+	}
+	if c.IsSet("complex1") {
+		if u, ok := any(&data.Complex1).(encoding.TextUnmarshaler); ok {
+			u.UnmarshalText([]byte(c.String("complex1")))
+		}
+	}
+	if c.IsSet("subtitle") {
+		emigo.ParseNullable(c.String("subtitle"), &data.Subtitle)
+	}
+	if c.IsSet("is-active") {
+		emigo.ParseNullable(c.String("is-active"), &data.IsActive)
+	}
+	if c.IsSet("is-featured") {
+		emigo.ParseNullable(c.String("is-featured"), &data.IsFeatured)
+	}
+	if c.IsSet("view-count") {
+		emigo.ParseNullable(c.String("view-count"), &data.ViewCount)
+	}
+	if c.IsSet("view-count-opt") {
+		emigo.ParseNullable(c.String("view-count-opt"), &data.ViewCountOpt)
+	}
+	if c.IsSet("small-count") {
+		emigo.ParseNullable(c.String("small-count"), &data.SmallCount)
+	}
+	if c.IsSet("small-count-opt") {
+		emigo.ParseNullable(c.String("small-count-opt"), &data.SmallCountOpt)
+	}
+	if c.IsSet("big-count") {
+		emigo.ParseNullable(c.String("big-count"), &data.BigCount)
+	}
+	if c.IsSet("big-count-opt") {
+		emigo.ParseNullable(c.String("big-count-opt"), &data.BigCountOpt)
+	}
+	if c.IsSet("ratio32") {
+		emigo.ParseNullable(c.String("ratio32"), &data.Ratio32)
+	}
+	if c.IsSet("ratio32-opt") {
+		emigo.ParseNullable(c.String("ratio32-opt"), &data.Ratio32Opt)
+	}
+	if c.IsSet("ratio64") {
+		emigo.ParseNullable(c.String("ratio64"), &data.Ratio64)
+	}
+	if c.IsSet("ratio64-opt") {
+		emigo.ParseNullable(c.String("ratio64-opt"), &data.Ratio64Opt)
+	}
+	if c.IsSet("status") {
+		emigo.ParseNullable(c.String("status"), &data.Status)
+	}
+	if c.IsSet("status-opt") {
+		emigo.ParseNullable(c.String("status-opt"), &data.StatusOpt)
+	}
+	if c.IsSet("metadata") {
+		emigo.ParseNullable(c.String("metadata"), &data.Metadata)
+	}
+	if c.IsSet("metadata-opt") {
+		emigo.ParseNullable(c.String("metadata-opt"), &data.MetadataOpt)
+	}
+	if c.IsSet("raw-settings") {
+		emigo.ParseNullable(c.String("raw-settings"), &data.RawSettings)
+	}
+	if c.IsSet("labels") {
+		emigo.ParseNullable(c.String("labels"), &data.Labels)
+	}
+	if c.IsSet("labels-opt") {
+		emigo.ParseNullable(c.String("labels-opt"), &data.LabelsOpt)
+	}
+	if c.IsSet("nested-container") {
+		emigo.ParseNullable(c.String("nested-container"), &data.NestedContainer)
+	}
+	if c.IsSet("nested-container-opt") {
+		emigo.ParseNullable(c.String("nested-container-opt"), &data.NestedContainerOpt)
+	}
+	return data
+}
+func GetEntity1EntityUpdateInputContent1CliFlags(prefix string) []emigo.CliFlag {
+	return []emigo.CliFlag{
+		{
+			Name: prefix + "item1",
+			Type: "int64?",
+		},
+	}
+}
+func CastEntity1EntityUpdateInputContent1FromCli(c emigo.CliCastable) Entity1EntityUpdateInputContent1 {
+	data := Entity1EntityUpdateInputContent1{}
+	if c.IsSet("item1") {
+		emigo.ParseNullable(c.String("item1"), &data.Item1)
+	}
+	return data
+}
+func GetEntity1EntityUpdateInputContent2CliFlags(prefix string) []emigo.CliFlag {
+	return []emigo.CliFlag{
+		{
+			Name: prefix + "item2",
+			Type: "int64?",
+		},
+	}
+}
+func CastEntity1EntityUpdateInputContent2FromCli(c emigo.CliCastable) Entity1EntityUpdateInputContent2 {
+	data := Entity1EntityUpdateInputContent2{}
+	if c.IsSet("item2") {
+		emigo.ParseNullable(c.String("item2"), &data.Item2)
+	}
+	return data
+}
+func GetEntity1EntityUpdateInputNestedContainerCliFlags(prefix string) []emigo.CliFlag {
+	return []emigo.CliFlag{
+		{
+			Name: prefix + "nested-inner",
+			Type: "object?",
+		},
+	}
+}
+func CastEntity1EntityUpdateInputNestedContainerFromCli(c emigo.CliCastable) Entity1EntityUpdateInputNestedContainer {
+	data := Entity1EntityUpdateInputNestedContainer{}
+	if c.IsSet("nested-inner") {
+		emigo.ParseNullable(c.String("nested-inner"), &data.NestedInner)
+	}
+	return data
+}
+func GetEntity1EntityUpdateInputNestedContainerNestedInnerCliFlags(prefix string) []emigo.CliFlag {
+	return []emigo.CliFlag{
+		{
+			Name: prefix + "nested-items",
+			Type: "complex",
+		},
+		{
+			Name: prefix + "nested-owner",
+			Type: "complex",
+		},
+	}
+}
+func CastEntity1EntityUpdateInputNestedContainerNestedInnerFromCli(c emigo.CliCastable) Entity1EntityUpdateInputNestedContainerNestedInner {
+	data := Entity1EntityUpdateInputNestedContainerNestedInner{}
+	if c.IsSet("nested-items") {
+		if u, ok := any(&data.NestedItems).(encoding.TextUnmarshaler); ok {
+			u.UnmarshalText([]byte(c.String("nested-items")))
+		}
+	}
+	if c.IsSet("nested-owner") {
+		if u, ok := any(&data.NestedOwner).(encoding.TextUnmarshaler); ok {
+			u.UnmarshalText([]byte(c.String("nested-owner")))
+		}
+	}
+	return data
+}
+func GetEntity1EntityUpdateInputNestedContainerOptCliFlags(prefix string) []emigo.CliFlag {
+	return []emigo.CliFlag{
+		{
+			Name: prefix + "nested-inner",
+			Type: "object?",
+		},
+	}
+}
+func CastEntity1EntityUpdateInputNestedContainerOptFromCli(c emigo.CliCastable) Entity1EntityUpdateInputNestedContainerOpt {
+	data := Entity1EntityUpdateInputNestedContainerOpt{}
+	if c.IsSet("nested-inner") {
+		emigo.ParseNullable(c.String("nested-inner"), &data.NestedInner)
+	}
+	return data
+}
+func GetEntity1EntityUpdateInputNestedContainerOptNestedInnerCliFlags(prefix string) []emigo.CliFlag {
+	return []emigo.CliFlag{
+		{
+			Name: prefix + "nested-items-opt",
+			Type: "complex",
+		},
+	}
+}
+func CastEntity1EntityUpdateInputNestedContainerOptNestedInnerFromCli(c emigo.CliCastable) Entity1EntityUpdateInputNestedContainerOptNestedInner {
+	data := Entity1EntityUpdateInputNestedContainerOptNestedInner{}
+	if c.IsSet("nested-items-opt") {
+		if u, ok := any(&data.NestedItemsOpt).(encoding.TextUnmarshaler); ok {
+			u.UnmarshalText([]byte(c.String("nested-items-opt")))
+		}
+	}
+	return data
+}
+
+// Entity1EntityCreateFn creates a new Entity1Entity row (and its array/collection/one relations,
+// including ones nested inside object/object? fields) from dto. dto.Id/dto.UniqueId are
+// assigned by the database (see AutoMigrate's column defaults) and populated back onto
+// dto once created. Relations are applied in a single transaction: one/one? are
+// resolved before the row itself is created (a belongs-to FK doesn't need the parent's
+// own id); array/array? and collection/collection? are reconciled afterwards, once
+// dto.Id is known.
+func Entity1EntityCreateFn(tx *gorm.DB, dto *Entity1Entity) (*Entity1Entity, error) {
+	err := tx.Transaction(func(tx *gorm.DB) error {
+		if dto.Owner.IsSet() {
+			var selectorId string
+			if dto.Owner.Operation == "select" {
+				if s, ok := dto.Owner.Selector.(string); ok {
+					selectorId = s
+				}
+			}
+			var item *Entity2Entity
+			if dto.Owner.Operation != "select" {
+				item = &dto.Owner.Item
+			}
+			resolvedId, err := emigorm.ReconcileOne(tx, dto.Owner.Operation, selectorId, item)
+			if err != nil {
+				return err
+			}
+			dto.OwnerId = resolvedId
+		}
+		if dto.Manager.IsSet() {
+			var selectorId string
+			if dto.Manager.Operation == "select" {
+				if s, ok := dto.Manager.Selector.(string); ok {
+					selectorId = s
+				}
+			}
+			var item *Entity2Entity
+			if dto.Manager.Operation != "select" {
+				item = &dto.Manager.Item
+			}
+			resolvedId, err := emigorm.ReconcileOne(tx, dto.Manager.Operation, selectorId, item)
+			if err != nil {
+				return err
+			}
+			dto.ManagerId = resolvedId
+		}
+		if v, ok := dto.Content2.Get(); ok && v != nil {
+			dto.Content2Row = v
+		}
+		if dto.NestedContainer.NestedInner.NestedOwner.IsSet() {
+			var selectorId string
+			if dto.NestedContainer.NestedInner.NestedOwner.Operation == "select" {
+				if s, ok := dto.NestedContainer.NestedInner.NestedOwner.Selector.(string); ok {
+					selectorId = s
+				}
+			}
+			var item *Entity2Entity
+			if dto.NestedContainer.NestedInner.NestedOwner.Operation != "select" {
+				item = &dto.NestedContainer.NestedInner.NestedOwner.Item
+			}
+			resolvedId, err := emigorm.ReconcileOne(tx, dto.NestedContainer.NestedInner.NestedOwner.Operation, selectorId, item)
+			if err != nil {
+				return err
+			}
+			dto.NestedContainer.NestedInner.NestedOwnerId = resolvedId
+		}
+		if v, ok := dto.NestedContainerOpt.Get(); ok && v != nil {
+			dto.NestedContainerOptRow = v
+		}
+		if err := tx.Create(dto).Error; err != nil {
+			return err
+		}
+		if dto.Items.IsSet() {
+			items := make([]*Entity1EntityItems, len(dto.Items.Items))
+			for i := range dto.Items.Items {
+				items[i] = &dto.Items.Items[i]
+			}
+			if err := emigorm.ReconcileHasMany(tx, "linker_id", dto.Id, dto.Items.Operation, items); err != nil {
+				return err
+			}
+		}
+		if dto.Items2.IsSet() {
+			items := make([]*Entity1EntityItems2, len(dto.Items2.Items))
+			for i := range dto.Items2.Items {
+				items[i] = &dto.Items2.Items[i]
+			}
+			if err := emigorm.ReconcileHasMany(tx, "linker_id", dto.Id, dto.Items2.Operation, items); err != nil {
+				return err
+			}
+		}
+		if dto.Items3.IsSet() {
+			items := make([]*Entity2Entity, len(dto.Items3.Items))
+			for i := range dto.Items3.Items {
+				items[i] = &dto.Items3.Items[i]
+			}
+			if err := emigorm.ReconcileManyToMany(tx, dto, "Items3Row", dto.Items3.Operation, items); err != nil {
+				return err
+			}
+		}
+		if dto.Items4.IsSet() {
+			items := make([]*Entity2Entity, len(dto.Items4.Items))
+			for i := range dto.Items4.Items {
+				items[i] = &dto.Items4.Items[i]
+			}
+			if err := emigorm.ReconcileManyToMany(tx, dto, "Items4Row", dto.Items4.Operation, items); err != nil {
+				return err
+			}
+		}
+		if dto.NestedContainer.NestedInner.NestedItems.IsSet() {
+			items := make([]*Entity1EntityNestedContainerNestedInnerNestedItems, len(dto.NestedContainer.NestedInner.NestedItems.Items))
+			for i := range dto.NestedContainer.NestedInner.NestedItems.Items {
+				items[i] = &dto.NestedContainer.NestedInner.NestedItems.Items[i]
+			}
+			if err := emigorm.ReconcileHasMany(tx, "linker_id", dto.Id, dto.NestedContainer.NestedInner.NestedItems.Operation, items); err != nil {
+				return err
+			}
+		}
+		if v, ok := dto.NestedContainerOpt.Get(); ok && v != nil {
+			if v.NestedInner.NestedItemsOpt.IsSet() {
+				items := make([]*Entity1EntityNestedContainerOptNestedInnerNestedItemsOpt, len(v.NestedInner.NestedItemsOpt.Items))
+				for i := range v.NestedInner.NestedItemsOpt.Items {
+					items[i] = &v.NestedInner.NestedItemsOpt.Items[i]
+				}
+				if err := emigorm.ReconcileHasMany(tx, "linker_id", dto.Id, v.NestedInner.NestedItemsOpt.Operation, items); err != nil {
+					return err
+				}
+			}
+		}
+		return nil
+	})
+	if err != nil {
+		return nil, err
+	}
+	return dto, nil
+}
+
+// Entity1EntityUpdateFn applies a partial update to the Entity1Entity row identified by id (its public
+// uniqueId, e.g. from an API path parameter - never the internal auto-increment id).
+// Only fields the caller actually set on input (input.{Field}.IsSet()) are touched -
+// anything else is left exactly as it was. one/one? are resolved into their {field}Id
+// FK column alongside the rest of the scalar changes; array/array? and
+// collection/collection? are reconciled afterwards via the same emigorm helpers
+// Entity1EntityCreateFn uses, against entity.Id (the row's real primary key, resolved from id
+// up front - gorm's Association API and the has-many reconcile both join on it, not on
+// uniqueId).
+func Entity1EntityUpdateFn(tx *gorm.DB, id string, input Entity1EntityUpdateInput) (*Entity1Entity, error) {
+	var entity Entity1Entity
+	err := tx.Transaction(func(tx *gorm.DB) error {
+		if err := tx.First(&entity, "unique_id = ?", id).Error; err != nil {
+			return err
+		}
+		changes := map[string]interface{}{}
+		if input.Owner.IsSet() {
+			var selectorId string
+			if input.Owner.Operation == "select" {
+				if s, ok := input.Owner.Selector.(string); ok {
+					selectorId = s
+				}
+			}
+			var item *Entity2Entity
+			if input.Owner.Operation != "select" {
+				item = &input.Owner.Item
+			}
+			resolvedId, err := emigorm.ReconcileOne(tx, input.Owner.Operation, selectorId, item)
+			if err != nil {
+				return err
+			}
+			changes["OwnerId"] = resolvedId
+		}
+		if input.Manager.IsSet() {
+			var selectorId string
+			if input.Manager.Operation == "select" {
+				if s, ok := input.Manager.Selector.(string); ok {
+					selectorId = s
+				}
+			}
+			var item *Entity2Entity
+			if input.Manager.Operation != "select" {
+				item = &input.Manager.Item
+			}
+			resolvedId, err := emigorm.ReconcileOne(tx, input.Manager.Operation, selectorId, item)
+			if err != nil {
+				return err
+			}
+			changes["ManagerId"] = resolvedId
+		}
+		if input.NestedContainer.IsSet() {
+			if v, ok := input.NestedContainer.Get(); ok && v != nil {
+				if v.NestedInner.IsSet() {
+					if v, ok := v.NestedInner.Get(); ok && v != nil {
+						if v.NestedOwner.IsSet() {
+							var selectorId string
+							if v.NestedOwner.Operation == "select" {
+								if s, ok := v.NestedOwner.Selector.(string); ok {
+									selectorId = s
+								}
+							}
+							var item *Entity2Entity
+							if v.NestedOwner.Operation != "select" {
+								item = &v.NestedOwner.Item
+							}
+							resolvedId, err := emigorm.ReconcileOne(tx, v.NestedOwner.Operation, selectorId, item)
+							if err != nil {
+								return err
+							}
+							changes["NestedOwnerId"] = resolvedId
+						}
+					}
+				}
+			}
+		}
+		if input.Title.IsSet() {
+			changes["Title"] = input.Title
+		}
+		if input.Content1.IsSet() {
+			if v, ok := input.Content1.Get(); ok && v != nil {
+				if v.Item1.IsSet() {
+					changes["Item1"] = v.Item1
+				}
+			}
+		}
+		if input.Content2.IsSet() {
+			if v, ok := input.Content2.Get(); ok && v != nil {
+				if v.Item2.IsSet() {
+					changes["Item2"] = v.Item2
+				}
+			}
+		}
+		if input.Complex1 != nil {
+			changes["Complex1"] = input.Complex1
+		}
+		if input.Subtitle.IsSet() {
+			changes["Subtitle"] = input.Subtitle
+		}
+		if input.IsActive.IsSet() {
+			changes["IsActive"] = input.IsActive
+		}
+		if input.IsFeatured.IsSet() {
+			changes["IsFeatured"] = input.IsFeatured
+		}
+		if input.ViewCount.IsSet() {
+			changes["ViewCount"] = input.ViewCount
+		}
+		if input.ViewCountOpt.IsSet() {
+			changes["ViewCountOpt"] = input.ViewCountOpt
+		}
+		if input.SmallCount.IsSet() {
+			changes["SmallCount"] = input.SmallCount
+		}
+		if input.SmallCountOpt.IsSet() {
+			changes["SmallCountOpt"] = input.SmallCountOpt
+		}
+		if input.BigCount.IsSet() {
+			changes["BigCount"] = input.BigCount
+		}
+		if input.BigCountOpt.IsSet() {
+			changes["BigCountOpt"] = input.BigCountOpt
+		}
+		if input.Ratio32.IsSet() {
+			changes["Ratio32"] = input.Ratio32
+		}
+		if input.Ratio32Opt.IsSet() {
+			changes["Ratio32Opt"] = input.Ratio32Opt
+		}
+		if input.Ratio64.IsSet() {
+			changes["Ratio64"] = input.Ratio64
+		}
+		if input.Ratio64Opt.IsSet() {
+			changes["Ratio64Opt"] = input.Ratio64Opt
+		}
+		if input.Status.IsSet() {
+			changes["Status"] = input.Status
+		}
+		if input.StatusOpt.IsSet() {
+			changes["StatusOpt"] = input.StatusOpt
+		}
+		if input.Metadata.IsSet() {
+			changes["Metadata"] = input.Metadata
+		}
+		if input.MetadataOpt.IsSet() {
+			changes["MetadataOpt"] = input.MetadataOpt
+		}
+		if input.RawSettings.IsSet() {
+			changes["RawSettings"] = input.RawSettings
+		}
+		if input.Labels.IsSet() {
+			changes["Labels"] = input.Labels
+		}
+		if input.LabelsOpt.IsSet() {
+			changes["LabelsOpt"] = input.LabelsOpt
+		}
+		if input.Misc != nil {
+			changes["Misc"] = input.Misc
+		}
+		if len(changes) > 0 {
+			if err := tx.Model(&entity).Updates(changes).Error; err != nil {
+				return err
+			}
+		}
+		if input.Items.IsSet() {
+			items := make([]*Entity1EntityItems, len(input.Items.Items))
+			for i := range input.Items.Items {
+				items[i] = &input.Items.Items[i]
+			}
+			if err := emigorm.ReconcileHasMany(tx, "linker_id", entity.Id, input.Items.Operation, items); err != nil {
+				return err
+			}
+		}
+		if input.Items2.IsSet() {
+			items := make([]*Entity1EntityItems2, len(input.Items2.Items))
+			for i := range input.Items2.Items {
+				items[i] = &input.Items2.Items[i]
+			}
+			if err := emigorm.ReconcileHasMany(tx, "linker_id", entity.Id, input.Items2.Operation, items); err != nil {
+				return err
+			}
+		}
+		if input.Items3.IsSet() {
+			items := make([]*Entity2Entity, len(input.Items3.Items))
+			for i := range input.Items3.Items {
+				items[i] = &input.Items3.Items[i]
+			}
+			if err := emigorm.ReconcileManyToMany(tx, &entity, "Items3Row", input.Items3.Operation, items); err != nil {
+				return err
+			}
+		}
+		if input.Items4.IsSet() {
+			items := make([]*Entity2Entity, len(input.Items4.Items))
+			for i := range input.Items4.Items {
+				items[i] = &input.Items4.Items[i]
+			}
+			if err := emigorm.ReconcileManyToMany(tx, &entity, "Items4Row", input.Items4.Operation, items); err != nil {
+				return err
+			}
+		}
+		if input.NestedContainer.IsSet() {
+			if v, ok := input.NestedContainer.Get(); ok && v != nil {
+				if v.NestedInner.IsSet() {
+					if v, ok := v.NestedInner.Get(); ok && v != nil {
+						if v.NestedItems.IsSet() {
+							items := make([]*Entity1EntityNestedContainerNestedInnerNestedItems, len(v.NestedItems.Items))
+							for i := range v.NestedItems.Items {
+								items[i] = &v.NestedItems.Items[i]
+							}
+							if err := emigorm.ReconcileHasMany(tx, "linker_id", entity.Id, v.NestedItems.Operation, items); err != nil {
+								return err
+							}
+						}
+					}
+				}
+			}
+		}
+		if input.NestedContainerOpt.IsSet() {
+			if v, ok := input.NestedContainerOpt.Get(); ok && v != nil {
+				if v.NestedInner.IsSet() {
+					if v, ok := v.NestedInner.Get(); ok && v != nil {
+						if v.NestedItemsOpt.IsSet() {
+							items := make([]*Entity1EntityNestedContainerOptNestedInnerNestedItemsOpt, len(v.NestedItemsOpt.Items))
+							for i := range v.NestedItemsOpt.Items {
+								items[i] = &v.NestedItemsOpt.Items[i]
+							}
+							if err := emigorm.ReconcileHasMany(tx, "linker_id", entity.Id, v.NestedItemsOpt.Operation, items); err != nil {
+								return err
+							}
+						}
+					}
+				}
+			}
+		}
+		return nil
+	})
+	if err != nil {
+		return nil, err
+	}
+	var updated Entity1Entity
+	if err := tx.First(&updated, "unique_id = ?", id).Error; err != nil {
+		return nil, err
+	}
+	return &updated, nil
+}
+
+// Entity1EntityActionsSig bundles the actions available for Entity1Entity. Extend this (and
+// Entity1EntityActions below) with more fields as more actions are generated - Create/Update
+// are wired to Entity1EntityCreateFn/Entity1EntityUpdateFn by default, but callers can swap either out
+// (e.g. in tests, or to layer extra validation/side effects around them).
+type Entity1EntityActionsSig struct {
+	Create func(tx *gorm.DB, dto *Entity1Entity) (*Entity1Entity, error)
+	Update func(tx *gorm.DB, id string, input Entity1EntityUpdateInput) (*Entity1Entity, error)
+}
+
+var Entity1EntityActions Entity1EntityActionsSig = Entity1EntityActionsSig{
+	Create: Entity1EntityCreateFn,
+	Update: Entity1EntityUpdateFn,
 }
